+++ cas-dotnet-sdk/PQC/SLHDSAWrapper.cs
@@ -20,11 +20,11 @@ namespace CasDotnetSdk.PQC
         {
             SLHDSAKeyPairStruct result = (this._platform == OSPlatform.Linux) ? SLHDSALinuxWrapper.slh_dsa_generate_signing_and_verification_key() : SLHDSAWindowsWrapper.slh_dsa_generate_signing_and_verification_key();
             byte[] signingKey = new byte[result.signing_key_length];
-            Marshal.Copy(result.signing_key_ptr, signingKey, 0, result.signing_key_length);
+            Marshal.Copy(result.signing_key_ptr, signingKey, 0, (int)result.signing_key_length);
             FreeMemoryHelper.FreeBytesMemory(result.signing_key_ptr);
 
             byte[] verificationKey = new byte[result.verification_key_length];
-            Marshal.Copy(result.verification_key_ptr, verificationKey, 0, result.verification_key_length);
+            Marshal.Copy(result.verification_key_ptr, verificationKey, 0, (int)result.verification_key_length);
             FreeMemoryHelper.FreeBytesMemory(result.verification_key_ptr);
             return new SLHDSAKeyPair
             {
@@ -39,7 +39,7 @@ namespace CasDotnetSdk.PQC
             SLHDSASignatureStruct result = (this._platform == OSPlatform.Linux) ? SLHDSALinuxWrapper.slh_dsa_sign_message(signingKey, signingKey.Length, message, message.Length) : SLHDSAWindowsWrapper.slh_dsa_sign_message(signingKey, signingKey.Length, message, message.Length);
             CasErrorHandler.ThrowIfError(result.error_code, "SLH-DSA sign");
             byte[] signature = new byte[result.signature_length];
-            Marshal.Copy(result.signature_ptr, signature, 0, result.signature_length);
+            Marshal.Copy(result.signature_ptr, signature, 0, (int)result.signature_length);
             FreeMemoryHelper.FreeBytesMemory(result.signature_ptr);
             return signature;
         }
